+++ src/main/java/tornadofx/Menu.kt
@@ -295,3 +295,15 @@ fun EventTarget.lazyContextmenu(op: ContextMenu.() -> Unit = {}) = apply {
         event.consume()
     }
 }
+
+fun ContextMenu.radiomenuitem(
+        name: String, toggleGroup: ToggleGroup? = null, keyCombination: KeyCombination? = null,
+        graphic: Node? = null, op: RadioMenuItem.() -> Unit = {}
+)  = RadioMenuItem(name, graphic).also {
+    toggleGroup?.apply { it.toggleGroup = this }
+    keyCombination?.apply { it.accelerator = this }
+    graphic?.apply { it.graphic = graphic }
+    op(it)
+    this += it
+}
+